lab4/requirements.txt
bcrypt==4.2.1
cffi==1.17.1
cryptography==44.0.1
future==1.0.0
jinja2==3.1.5
MarkupSafe==2.1.5
netmiko==4.4.0
ntc-templates==7.7.0
paramiko==3.5.1
pycparser==2.22
PyNaCl==1.5.0
pyserial==3.5
PyYAML==6.0.2
scp==0.15.0
six==1.17.0
textfsm==1.1.3
